refactor: replace wirteResult ro abstract class
- Перенес код, который пишет статистику в отдельный абстракстный класс
